requirements.txt
flask
requests
mysql-connector-python
google-cloud-secret-manager